LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/text-decoration-skip-expected.txt


FAIL Can set 'text-decoration-skip' to CSS-wide keywords: initial assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'text-decoration-skip' to CSS-wide keywords: inherit assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'text-decoration-skip' to CSS-wide keywords: unset assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'text-decoration-skip' to CSS-wide keywords: revert assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'text-decoration-skip' to var() references:  var(--A) assert_equals: expected 2 but got 1
FAIL Can set 'text-decoration-skip' to the 'none' keyword: none assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'text-decoration-skip' to the 'objects' keyword: objects Bad value for shorthand CSS property
FAIL Can set 'text-decoration-skip' to the 'edges' keyword: edges Bad value for shorthand CSS property
FAIL Can set 'text-decoration-skip' to the 'box-decoration' keyword: box-decoration Bad value for shorthand CSS property
FAIL Can set 'text-decoration-skip' to the 'spaces' keyword: spaces Bad value for shorthand CSS property
PASS Setting 'text-decoration-skip' to a length throws TypeError
PASS Setting 'text-decoration-skip' to a percent throws TypeError
PASS Setting 'text-decoration-skip' to a time throws TypeError
PASS Setting 'text-decoration-skip' to an angle throws TypeError
PASS Setting 'text-decoration-skip' to a flexible length throws TypeError
PASS Setting 'text-decoration-skip' to a number throws TypeError
PASS Setting 'text-decoration-skip' to a URL throws TypeError
PASS Setting 'text-decoration-skip' to a transform throws TypeError
FAIL 'text-decoration-skip' does not supported 'objects spaces' assert_not_equals: Unsupported value must not be null got disallowed value null
FAIL 'text-decoration-skip' does not supported 'leading-spaces trailing-spaces' assert_not_equals: Unsupported value must not be null got disallowed value null
FAIL 'text-decoration-skip' does not supported 'objects edges box-decoration' assert_not_equals: Unsupported value must not be null got disallowed value null
